bugs fixing

+++ public/js/applicantModal.js
@@ -16,7 +16,7 @@ function prettyDate(date) {
         return null;
     }
     date = new Date(date);
-    let dateOptions = {year: 'numeric', month: 'numeric', day: 'numeric'};
+    let dateOptions = {year: 'numeric', month: 'long', day: 'numeric'};
     return date.toLocaleDateString("en-GB", dateOptions)
 
 }
